requirements.txt
tqdm==4.66.4
matplotlib==3.8.0
scipy==1.11.4
gdown==5.2.0
scikit-image==0.22.0
numpy==1.22.3
chumpy==0.70
tensorboard==2.12.3
fvcore==0.1.5.post20221221
iopath==0.1.10
plyfile==1.0.3
tyro==0.8.5
dearpygui==1.11.1
roma==1.5.0
trimesh==4.0.0
transformers==4.41.2
diffusers==0.28.1
einops==0.8.0
ffmpeg-python==0.2.0
imageio==2.34.2
imageio-ffmpeg==0.4.8
lpips==0.1.4
opencv-python==4.9.0.80
pillow==10.4.0
protobuf==4.25.3
PyYAML==6.0.1
timm==0.4.12
xformers==0.0.26.post1
smplx==0.1.28
click==8.1.7
mrcfile==1.5.0
pyrender==0.1.45
PyOpenGL==3.1.0 
ninja==1.11.1.1
huggingface-hub==0.25.0
accelerate==0.30.1
omegaconf==2.3.0
onnxruntime==1.18.0
onnxruntime-gpu==1.18.0
datasets==2.19.2
wandb==0.17.0 
peft==0.11.1
pymeshlab==2023.12.post1
xatlas==0.0.9
git+https://github.com/NVlabs/nvdiffrast
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch